:memo: landing 페이지 업데이트
