--- src/components/GlassSearch/Landing.module.scss
@@ -1,114 +1,0 @@
-.loadingPg {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 100vh;
-    font-size: 26px;
-    animation: pulse 0.4s 3;
-    color: #6b799e;
-    font-weight: 700;
-
-  }
-  
-  @keyframes pulse {
-    0% {
-      font-size: 26px;
-    }
-    50% {
-      font-size: 32px;
-    }
-    100% {
-      font-size: 26px;
-    }
-  }
-  
-  
-
-.weatherbox {
-    align-items: center;
-    text-align: center;
-    border-radius: 5px;
-    margin: 10px;
-    width: 700px;
-    border: dotted 1px black;
-    padding: 10px;
-    // background-color: #f6f9fa;
-
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-
-    line-height: 15px;
-
-}
-
-.drinkImg {
-    width: 350px;
-    height: 350px;
-    border-radius: 5px;
-    transition: all 0.3s ease;
-
-    &:hover {
-        transform: scale(0.98);
-    }
-}
-
-
-// 날씨 아이콘 시작
-.weatherIcons {
-  overflow: hidden;
-  width: 20%;
-  display: flex;
-}
-
-.weatherIcon {
-  animation: moveX 3.5s infinite linear;
-}
-
-@keyframes moveX {
-  0% {
-    transform: translateX(-50%); 
-  }
-  100% {
-    transform: translateX(150%); 
-  }
-}
-// 날씨 아이콘 끝
-
-
-// 버튼 박스 시작
-.buttonBox {
-  display: flex;
-  justify-content: space-between;  // 버튼 사이에 공간 추가
-}
-
-.learnBtn, .againBtn {
-  background-color: #6b799e;
-  border: none;
-  color: white;
-  cursor: pointer;
-  line-height: 0.5;
-  border-radius: 5px;
-  margin: 10px 5px 15px 0;
-  padding: 15px;
-  text-align: center;
-  font-weight: 600;
-  transition: box-shadow 400ms ease-in-out, color 300ms ease-in-out;
-
-  &:focus {
-    color: white;
-    outline: 0;
-  }
-  &:hover {
-    box-shadow: 0 0 40px 40px #ebc57c inset;
-    color: white;
-    border: none;
-  }
-}
-// 버튼 박스 끝
-
-.cocktailName {
-    color: #9c8f96;
-    font-size: 26px;
-    margin: 20px auto;
-}
